8.15 done

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async(req, res) => {
     const videos = await Video.find({}).sort({createdAt:"desc"});
@@ -7,7 +8,9 @@ export const home = async(req, res) => {
 
 export const see = async(req, res)=>  {
   const {id} =req.params; //same req.params.id 
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  // const videoOwner = await User.findById(video.owner);
+  // console.log(video);
   if(!video){
     return res.status(404).render("404", {pageTitle:"Video not found."})
   }
@@ -16,7 +19,13 @@ export const see = async(req, res)=>  {
 
 export const getEdit = async(req, res) =>  {
     const {id} =req.params; 
+    const {
+      user:{_id}
+    } =req.session;
     const video = await Video.findById(id);
+    if(String(video.owner) !== String(_id)){
+      return res.status(403).redirect("/");
+    }
     if(!video){
       return res.render("404", {pageTitle:"Video not found."})
     }
@@ -25,10 +34,16 @@ export const getEdit = async(req, res) =>  {
 
 export const postEdit = async(req, res) =>{ 
   const {id} =req.params;
+  const {
+    user:{_id}
+  } =req.session;
   const {title, description, hashtags} =req.body; //form에서 name에 전송된 정보
 
 const video = await Video.exists({_id:id});
-    if(!video){
+if(String(video.owner) !== String(_id)){
+  return res.status(403).redirect("/");
+}
+if(!video){
       return res.status(404).render("404", {pageTitle:"Video not found."})
     }
   //  const video = await Video.findById(id);
@@ -59,16 +74,23 @@ export const getUpload = (req, res) =>{
   return res.render("upload",{pageTitle:"Upload Video"})};
 
 export const postUpload = async (req, res) => {
+  const {
+    user:{_id}
+  } = req.session;
     const { title, description, hashtags } = req.body;
     const {file} = req;
     console.log(req.file);
     try{
-      await Video.create({
+      const newVideo = await Video.create({
         title:title,
         fileUrl:file.path,
+        owner:_id,
         description:description,
         hashtags:Video.formatHashtags(hashtags),
       });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/")
     } catch(error){
       return res.status(400).render("upload",{pageTitle:"Upload Video", errorMessage:error._message,})
@@ -77,7 +99,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req,res)=>{
  const {id} =req.params;
-  await Video.findByIdAndDelete(id);
+ const {user:{_id}} =req.session;
+
+ const video = await Video.findById(id);
+ if (!video) {
+  return res.status(404).render("404", { pageTitle: "Video not found." });
+}
+if(String(video.owner) !== String(_id)){
+  return res.status(403).redirect("/");
+}
+
+await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
